Undo changes to CollectionService
